requirements.txt
requests
python-dotenv
PySimpleGUI
pandas
beautifulsoup4
